added another test for spying, but separated some independent functions out into a utils file to play with mocking later
